my version of account

+++ src/main/java/com/priceminister/account/implementation/CustomerAccountRule.java
@@ -16,14 +16,13 @@ package com.priceminister.account.implementation;
 
 import com.priceminister.account.*;
 
+import java.math.BigDecimal;
+
 
 public class CustomerAccountRule implements AccountRule {
 
-    /* (non-Javadoc)
-     * @see com.priceminister.account.AccountRule#withdrawPermitted(java.lang.Double)
-     */
-    public boolean withdrawPermitted(Double resultingAccountBalance) {
-        return resultingAccountBalance >= 0;
+    public boolean withdrawPermitted(BigDecimal resultingAccountBalance) {
+        return resultingAccountBalance.doubleValue() >= 0;
     }
 
 }
